clean up comments on scripts and move execShellCommand to shellHelper.ts
