+++ src/scripts/cloneRepo.sh
@@ -10,8 +10,12 @@ temp_sock=$(cat ./tmpAgent/agentSock)
 # add the ssh agent sock to the environment variables
 export SSH_AUTH_SOCK=$temp_sock
 
-# create a myRepos folder to store repositories
+# create a myRepos folder to store repositories if it doesn't already exist
+if [ ! -d "./myRepos" ] 
+then
 mkdir ./myRepos
+fi
+
 cd myRepos
 
 # clone a git repository from github using ssh connection
